% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faviconPlease.R
\name{faviconDuckDuckGo}
\alias{faviconDuckDuckGo}
\title{Use DuckDuckGo's favicon service}
\usage{
faviconDuckDuckGo(server)
}
\arguments{
\item{server}{The name of the server, e.g. "www.r-project.org"}
}
\value{
Character vector
}
\description{
The search engine \href{https://duckduckgo.com/}{DuckDuckGo} includes site
favicons in its search results, and it makes this service publicly available.
If it can't find a favicon, it returns a default fallback. faviconPlease uses
this as a fallback function if the favicon can't be found directly via the
standard methods.
}
\examples{
  faviconDuckDuckGo("reactome.org")

}
\references{
\href{https://duckduckgo.com/duckduckgo-help-pages/privacy/favicons/}{DuckDuckGo favicons privacy}
}
\seealso{
\code{\link{faviconPlease}},
         \code{\link{faviconGoogle}}
}
