% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faviconPlease.R
\name{faviconGoogle}
\alias{faviconGoogle}
\title{Use Google's favicon service}
\usage{
faviconGoogle(server)
}
\arguments{
\item{server}{The name of the server, e.g. "www.r-project.org"}
}
\value{
Character vector
}
\description{
Use Google's favicon service
}
\examples{
  faviconGoogle("reactome.org")

}
