% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faviconPlease.R
\name{faviconIco}
\alias{faviconIco}
\title{Check for the existence of favicon.ico}
\usage{
faviconIco(scheme, server, path)
}
\arguments{
\item{scheme}{"http" or "https"}

\item{server}{The name of the server, e.g. "www.r-project.org"}

\item{path}{The path to a target file on the server (must start with a
forward slash)}
}
\description{
Check for the existence of favicon.ico
}
