% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faviconPlease.R
\name{faviconLink}
\alias{faviconLink}
\title{Search for a link element that specifies the location of the favicon}
\usage{
faviconLink(scheme, server, path)
}
\arguments{
\item{scheme}{"http" or "https"}

\item{server}{The name of the server, e.g. "www.r-project.org"}

\item{path}{The path to a target file on the server (must start with a
forward slash)}
}
\description{
Search for a link element that specifies the location of the favicon
}
