% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faviconPlease.R
\name{faviconPlease}
\alias{faviconPlease}
\title{Find the URL to a website's favicon}
\usage{
faviconPlease(
  links,
  functions = list(faviconLink, faviconIco),
  fallback = faviconDuckDuckGo
)
}
\description{
Find the URL to a website's favicon
}
